Refactor integer handling to use qint64 instead of int
- Updated various functions in lexicons_func.cpp to replace `toInt()` with `toInteger()` for better handling of large integers.
- Modified defs2struct.py to change integer type definitions from `int` to `qint64` for consistency and to accommodate larger values.
- Ensured that all instances of integer handling in JSON parsing and struct definitions are updated accordingly.

+++ lib/atprotocol/app/bsky/actor/appbskyactorsearchactors.cpp
@@ -9,7 +9,7 @@ AppBskyActorSearchActors::AppBskyActorSearchActors(QObject *parent)
     m_listKey = QStringLiteral("actors");
 }
 
-void AppBskyActorSearchActors::searchActors(const QString &q, const int limit,
+void AppBskyActorSearchActors::searchActors(const QString &q, const qint64 limit,
                                             const QString &cursor)
 {
     QUrlQuery url_query;
