+++ lib/atprotocol/app/bsky/feed/appbskyfeedgetlistfeed.h
@@ -10,7 +10,7 @@ class AppBskyFeedGetListFeed : public AppBskyFeedGetTimeline
 public:
     explicit AppBskyFeedGetListFeed(QObject *parent = nullptr);
 
-    void getListFeed(const QString &list, const int limit, const QString &cursor);
+    void getListFeed(const QString &list, const qint64 limit, const QString &cursor);
 };
 
 }
